server: Refactor code block theme search into own service
